==> 1_data_types.ipynb <==
# Notebook 1

## 01. Import libraries

# Import libraries
import pandas as pd
import numpy as np
import os

## 02. Working with Python data types

### Step 7

x = 2
y = 5
z = 10

print (x * y * z )

print(y * z * 2)

print(z * 10 )

### Step 8

a = 1.234

type (a)

b = 2.345

type (b)

print(a/b)

### Step 9 

part1 = "Py"

part2 = "thon"

part1 + part2 

### Step 10

#first sentence 
part3 = "I "
part4 = "love "
part5 = "my "
part6 = "dog"

sentence1 = part3+part4+part5+part6

print (sentence1)

#second sentence
part7 = ", her "
part8 = "name "
part9 = "is "
part10 = "Hope. "

sentence2 = part7+part8+part9+part10

print(sentence2)

#short word
print (sentence1+sentence2)

### Bonus task

n = 'Hope ' 
print(n*100)
