==> 2_pandas.ipynb <==
# 01. Importing Libraries

### Step 3

#Import libraries
import pandas as pd
import numpy as np
import os

### Step 4

df = pd.read_csv(...\02 Data\Original Data\orders.csv', index_col = False);

path = r'C:\Users\Sthif\01-2025-Instacart Basket Analysis'

### Step 5

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'Orders.csv'), index_col = False)

df.columns

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'Orders.csv'), usecols = vars_list)

### Step 6

df.head()

### Step 7

df_prods = pd.read_csv(r'C:\Users\Sthif\01-2025-Instacart Basket Analysis\02 Data\Original Data\products.csv', index_col = False)

path = r'C:\Users\Sthif\01-2025-Instacart Basket Analysis'

df_prods=pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'Products.csv'), index_col = False)

### Step 8

df_prods.head(20)

df_prods.tail(35)

### Step 9

df_prods.columns

### Step 10

df_prods.info()

### Step 11

df_prods.describe()

#Max value for aisle_id is 134.000000

df_prods.dtypes

#Data type for department_id is int64
