==> 4_data_consistency_checks.ipynb <==
### Step 1:

# Import libraries
import pandas as pd
import numpy as np
import os

df_prods = pd.read_csv(...\02 Data\Original Data\products.csv', index_col = False)
path = ...'
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_ords = pd.read_csv(...\02 Data\Prepared Data\orders_wrangled.csv', index_col = False)
path = ...'
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# Data Consistency
df_ords.describe()

# Create a DataFrame
df_test = pd.DataFrame()
# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]
df_test.head()

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

df_test['mix'] = df_test['mix'].astype('str')

df_test['mix'] = df_test['mix'].astype('str')
df_prods.isnull().sum()
df_nan = df_prods [df_prods['product_name'].isnull() == True]
df_nan

# Missing Values
df_prods.shape

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]
df_prods_clean.shape

# Finding Duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

# Another way to drop missing values
df_prods.dropna(inplace = True)
# And for a particular column, only 
df_prods.dropna(subset = ['product_name'], inplace = True)

# Clean duplicates
df_prods_clean.shape

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups.shape

### Step 2:

df_ords.describe()

#### Brief Analysis of df.describe() Output
##### 1.order_number: the maximum value of 100 seems unusually high and shoudl be investigated to confirm validity.
##### 2.days_prior_order: Missing values are present, end rows with a value of 0 shoudl be verified to ensure they 
##### represent valid reorders.

### Step 3:

# Check for mix-type data
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

df_ords.info()

### Step 4, 5 & 6:

# No mixed-type data
df_ords.isnull().sum()

print("Missing values in df_ords:")

# No missing values
df_ords.shape

### Step 7 & 8:

df_dups = df_ords[df_ords.duplicated()]
print(df_dups)

# Dropping the "unnamed:0" column
df_ords = df_ords.drop(columns=['Unnamed: 0'])

df_ords.describe()

df_ords.head()

### Step 9:

# Export your final, cleaned df_prods and df_ords
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))
df_prods_clean.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_wrangled.csv'))
