==> 5_combining_and_exporting_data_part1.ipynb <==
### Import libraries & DF

#Import libraries:
import pandas as pd
import numpy as np
import os

df_ords_prior = pd.read_csv(...\02 Data\Original Data\orders_products_prior.csv', index_col = False)
path = ...
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

df_ords = pd.read_csv(...\02 Data\Prepared Data\orders_wrangled.csv', index_col = False)
path = ...'
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# Check the output
df_ords_prior.head()

# Check the output
df_ords = df_ords.drop(columns=['Unnamed: 0'])
df_ords.head()

df_ords_prior.shape

df_ords.shape

# Merge DataFrames
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

# Export Data in Pickle Format
df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
