==> 5_combining_and_exporting_data_part2.ipynb <==
### Import libraries:

#Import libraries:
import pandas as pd
import numpy as np
import os

#import the orders_products_combined
orders_products_combined = pd.read_pickle(...\02 Data\Prepared Data\orders_products_combined.pkl')
path = ...'
orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

### Step 4:

#Check the shape of the imported dataframe 
orders_products_combined.shape

### Step 5:

#Determine a suitable way to combine the orders_products_combined
df_prods = pd.read_csv(...\02 Data\Prepared Data\products_wrangled.csv', index_col = False)
path = ...'
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_wrangled.csv'), index_col = False)

### Step 6:

#Confirm the results of the merge using the merge flag.
orders_products_combined.head()

df_prods.head()

### Step 7:

print(orders_products_combined.shape)
print(df_prods.shape)

pd.merge(orders_products_combined , df_prods , on = ['product_id'] , how = 'inner', indicator = 'match' )

# Dropping the "unnamed:0" column
df_prods = df_prods.drop(columns=['Unnamed: 0'])

# Merge of 'orders_products_combined' and 'df_prods'
ords_prods_merge = orders_products_combined.merge(df_prods , on = ['product_id'] , indicator = 'match' )

#Export the newly created dataframe as ords_prods_merge in a suitable format
# Dropping the "_merge and match" column
ords_prods_merge = ords_prods_merge.drop(columns=['_merge', 'match'])

# Export Data in Pickle Format
ords_prods_merge.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
