==> 8_data_visualization_part1.ipynb <==
### Step 3:

# Import libraries
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

#import the orders_products_combined
ords_prods_merge = pd.read_pickle(...\02 Data\Prepared Data\orders_products_combined.pkl')
path = ...'
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

#import the customer data
customers = pd.read_csv(...\02 Data\Original Data\customers.csv')
path = ...'
customers = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

### Step 4 & 5:

# Wrangling customers dataset to ensure they are clean, logical, and consistent for analysis.

# 1. View the structure of the customers dataset
customers.head()  # Display the first few rows of the dataset
customers.info()  # Show column names, data types, and null values


# 2. Check for missing values
# Identify if any columns contain missing values and determine how to handle them (e.g., fill or drop).
customers.isnull().sum()  # Check for missing values in each column

# 3. Check for duplicate rows
customers.duplicated().sum()  # Count duplicate rows

# 4. Rename columns for clarity
customers.rename(columns={'First Name': 'first_name'}, inplace=True)
customers.rename(columns={'Surnam': 'last_name'}, inplace=True)
customers.rename(columns={'Gender': 'gender'}, inplace=True)
customers.rename(columns={'STATE': 'state'}, inplace=True)
customers.rename(columns={'Age': 'age'}, inplace=True)
customers.rename(columns={'n_dependants': 'num_dependants'}, inplace=True)
customers.rename(columns={'fam_status': 'civil_status'}, inplace=True)

# Replace NaN values in 'first_name' with a placeholder
customers['first_name'] = customers['first_name'].fillna('Unknown')

customers.info()

# 5. Check for mixed data types in each column
for col in customers.columns:
    print(f"{col}: {customers[col].apply(type).value_counts()}")

### Step 6:

# 1. Check the data
ords_prods_merge.head()
ords_prods_merge.info()

# 2. Merge ords_prods_merge with customers on the 'user_id' column
df_merged = pd.merge(ords_prods_merge, customers, on='user_id', how='left', indicator=True)

# 3. Drop the '_merge' column
df_merged.drop(columns=['_merge'], inplace=True)

df_merged.info()

# Export the cleaned DataFrame to a pickle file for further use
df_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'instacart_data.pkl'))
